Enable Tree-Structured Parzen Estimator (TPE) Strategy
Sequential model-based optimization methods (SMBO) are a formalization of Bayesian optimization.
The sequential refers to running trials one after another, each time trying better hyperparameters
by applying Bayesian reasoning and updating a surrogate model.

There are five main aspects of SMBO:
1. Domain: A domain of hyperparameters or search space
2. Objective function: An objective function which takes hyperparameters as input and outputs
a score that needs to minimize (or maximize)
3. Surrogate function: The representative surrogate model of the objective function
4. Selection function: A selection criterion to evaluate which hyperparameters to choose next
trial from the surrogate model
5. History: A history consisting of (score, hyperparameter) pairs used by the algorithm to
update the surrogate model

There are several variants of SMBO, which differ in how to build a surrogate and the selection
criteria (steps 3–4). The TPE builds a surrogate model by applying Bayes's rule.

**NOTE**:
TPE requires many iterations to converge to an optimal solution, and it is recommended to run
it for at least 200 iterations. Because every iteration requires evaluation of a generated model,
which means accuracy measurements on a dataset and latency measurements using a benchmark, this
process may take from 24 hours up to few days to complete, depending on a model.