[SW-233498] Fix scale calculation cguid flow in all scale methods (#263)
Enabling scale calculation CGUID for static quantization, when calculating weight scale.
A flow that doesn't pass in CGUID requires to divide maxabs in fullscale and backoff factor
In PTS there's a cast to hp_dtype.
Note: In test_qdq there was a type mismatch that required explicitly casting to hp_dtype in CGUID call

Co-authored-by: linoy buchnik <lbuchnik@habana.ai>