[SW-220805] Fixing zero values with FSDPA (#252)
In the input, we use zero tokens for padding. After the linear layer, we set the corresponding positions (from the padding) to -inf, so that the softmax outputs values close to epsilon.

When using FSDPA optimization, to improve performance, we avoid copying the -inf values to the softmax and instead set them directly to zero. As a result, the softmax output becomes exactly zero (as opposed to a small epsilon value without the FSDP optimization).

When computing the dynamic scale for the out_proj, this leads to a division by zero issue.

The fix we're implementing is to use max(epsilon, scale) during scale calc.

This fix aligns non-CGUID code to act the same as the CGUID flow